+++ miataru/miataru/views/iPhone/iPhone_DevicesView.swift
@@ -63,7 +63,7 @@ struct iPhone_DevicesView: View {
                             )
                             .swipeActions(edge: .leading, allowsFullSwipe: false) {
                                 Button {
-                                    appNavigation.openAddDevice(visitor.DeviceID)
+                                    appNavigation.openAddDevice(visitor.DeviceID, source: .unknownVisitor)
                                 } label: {
                                     Label(
                                         settings.allowedDeviceListEnabled